Added Android navigation bar back :(
Encountered an issue where leaving the app and coming back would respawn the navigation bar, leading to less than desirable results for pages that rely on the screen dimensions being consistent.

I added the navigation bar back for now. Hopefully there will be a fix for this in Expo sometime in the future.